# roadsos-backend/docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: RoadSoS API
  version: 1.0.0
servers:
  - url: http://localhost:5000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/admin/queue/{name}/retry:
    post:
      parameters: [{ in: path, name: name, required: true, schema: { type: string } }]
      responses: { '200': { description: Failed jobs retried } }